Fixed A Lot, Added New pages ETC, Added Error Pages, Auth handleing

+++ Backend/routes/notesGroup.js
@@ -2,9 +2,10 @@ const router = require("express").Router();
 const shortuid = require("short-unique-id");
 const idgen = new shortuid({ length: 12 });
 const { notesGroupCol } = require("../db/dbconnection");
+const { deleteAllOfGroup } = require("./notes");
 
 router.get("/getAll", async (req, res) => {
-  var cur = await notesGroupCol.find({ ownerID:  req.user.loggedinUserUUID });
+  var cur = await notesGroupCol.find({ ownerID: req.user.loggedinUserUUID });
   res.send(await cur.toArray());
 });
 
@@ -24,15 +25,16 @@ router.post("/update", async (req, res) => {
 
 router.post("/new", async (req, res) => {
   if (req.body.title && req.body.description) {
+    id = idgen.rnd();
     notesGroupCol
       .insertOne({
         ownerID: req.user.loggedinUserUUID,
-        groupID: idgen.rnd(),
+        groupID: id,
         title: req.body.title,
         description: req.body.description,
         favourite: false,
       })
-      .then(() => res.sendStatus(200))
+      .then(() => res.status(200).send(id))
       .catch(() => res.sendStatus(500));
   } else {
     res.sendStatus(400);
@@ -40,24 +42,25 @@ router.post("/new", async (req, res) => {
 });
 
 router.post("/editFavourite", async (req, res) => {
-    notesGroupCol
-      .updateOne(
-        { ownerID: req.user.loggedinUserUUID, groupID: req.body.id },
-        { $set: { favourite : req.body.favStatus } }
-      )
-      .then(() => res.sendStatus(200))
-      .catch(() => res.sendStatus(500));
-  } 
-);
-
-router.delete("/deleteNoteGroup" , async (req, res) => {
   notesGroupCol
-    .deleteOne(
-      { ownerID: req.user.loggedinUserUUID, groupID: req.body.id }
+    .updateOne(
+      { ownerID: req.user.loggedinUserUUID, groupID: req.body.id },
+      { $set: { favourite: req.body.favStatus } }
     )
     .then(() => res.sendStatus(200))
     .catch(() => res.sendStatus(500));
-} 
-);
+});
+
+router.delete("/deleteNoteGroup", async (req, res) => {
+
+  notesGroupCol
+    .deleteOne({ ownerID: req.user.loggedinUserUUID, groupID: req.body.id })
+    .then(async () => {
+      var resl = await deleteAllOfGroup(req.user.loggedinUserUUID,req.body.id);
+      if (resl) res.sendStatus(200);
+      else res.sendStatus(500);
+    })
+    .catch(() => res.sendStatus(500));
+});
 
 module.exports = router;
